cloud-service-api/requirements/requirements.txt
click==8.1.2
flask_sqlalchemy
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1
marshmallow_sqlalchemy
gunicorn
pyjwt
